SERVER-6407 Use saslauthd for SASL PLAIN password validation on the $external database.
This change, plus proper configuration of saslauthd allows administrators to
configure PAM and LDAP (and other) proxy authentication mechanisms.  It also
provides a route to limited kerberos proxy authentication, for users unable to
use Kerberos support in the standard way.

Its primary limitation is that the SASL PLAIN mechanism involves transmitting
the user password to the mongo server, so should generally only be used on a
trusted channel (VPN, SSL, trusted wired network).

The patch adds one new server parameter, saslauthdPath, which is the path to the Unix
Domain Socket of the saslauthd instance to use for proxy authentication.

The patch stops using the user_realm to convey the authentication database to
the auxprop_mongodb_internal module, instead taking advantage of the fact that
we now pass the SaslAuthenticationSession object to that module via a special
sasl_callback_t.  This was necessary because saslauthd's LDAP mode passes the
user realm value on to saslauthd, and through it to the ldap server, stymying
authentication.

To use this authentication mechanism in the shell, it is necessary to pass
digestPassword: false to db.auth() when authenticating on the $external
database, since the server must receive an undigested password to forward on to
saslauthd.

To use with ActiveDirectory, saslauthd should be started with the following
config file contents:

ldap_servers: <ldap uri, e.g. ldaps://ad.mydomain.com>
ldap_use_sasl: yes
ldap_mech: DIGEST-MD5
ldap_auth_method: fastbind

To configure the mongo server, if saslauthd has a socket path of /foo/bar/saslauthd, set the
saslauthdPath parameter to /foo/bar/saslauthd/mux.  Leave the value set to the empty string to use
the library default value.