WRITING-2730 Introduce mongoMain.
This is needed to allow the mql executable to build without
explicitly listing all of the modules inside the mql-model
library. This is prep work for beginning to implement the
mql command line tool.

This commit also introduces a dependency on the
optparse-applicative package, which we will use for parsing
command-line arguments.