SERVER-40477 Make BSON type set available in schema metadata tree.
This is precursor work which will be built upon in a
subsequent patch. By making the BSON type information
available in the schema tree, future code can validate that
comparisons and writes to encrypted fields are sensible with
respect to BSON type, and can return an error to the client
in type mismatch scenarios.

Prior to this change, we were overloading a single type,
EncryptionMetadata, in order to represent both the parsed
version of what users can specify inside the
'encryptMetadata' keyword and to represent the resolution of
the 'encryptMetadata' chain. This patch separates those
concepts into distinct types called EncryptionMetadata and
ResolvedEncryptionInfo respectively.

The resolved encryption info differs from the encryption
metadata variant in that the 'algorithm' and 'keyId' fields
are non-optional. Furthermore, the resolved metadata, unlike
the heritable metadata, may contain a set of BSON types.
Finally, EncryptionMetadata is an IDL type that has a
serialization to and from BSON, whereas
ResolvedEncryptionInfo has no BSON representation.