SERVER-41253 Add support for mongotTaskExecutor
Now that mongot no longer needs a non-authing executor, it needs a
regular executor without hooks it can use.

Given that we don't have one of those, let's install one, specifically
for it's use.

At some point in the future we'll have to revisit how many of these we
make and use.