SERVER-41264 Make mongocryptd error on attempt to randomly encrypt single-valued type.
Prior to this change, null, undefined, minKey, and maxKey
were banned if specified under 'encrypt.bsonType'. However,
with the random encryption algorithm, 'encrypt.bsonType' is
not required. If the client's schema specified random
encryption without any type information, mongocryptd could
incorrectly mark an illegal type for random encryption. That
flaw is corrected in this patch, by checking for an illegal
type in two places -- first, when analyzing the schema, and
second, when marking a BSON element for encryption.